[FEATURE] Add ability to post (without image)
The user can click a button on the home screen that will open up a camera. The user can then take a picture but the picture does not save (that is on the TODO list). After that, an intent is fired that takes the user to the "create post" activity where they are able to add a description. After adding a description, the user can click a post button to post the post to the parse server.